[#74] Feat(useAuth): useAuth 커스텀 훅 적용
